+++ packages/elements/src/core/bind/update-attributes.ts
@@ -1,4 +1,5 @@
 import type { AttributeMaps } from "../../types/index.js";
+import C from "../constants.js";
 import Elements from "../elements.js";
 import scope from "../scope/index.js";
 import {
@@ -58,10 +59,12 @@ const updateAttributes = (
 	attributeMaps: AttributeMaps | undefined,
 ) => {
 	if (!attributeMaps?.bindState) return;
+	//* prefix it with the state key $ so it matches that in the bindState map
+	const stateWithPrefix = `${C.defaults.attributes.denote.state}${state.key}`;
 
 	const stateKey = attributeMaps.scope
-		? scope.scopeValue(attributeMaps.scope, state.key)
-		: state.key;
+		? scope.scopeValue(attributeMaps.scope, stateWithPrefix)
+		: stateWithPrefix;
 
 	const affectedAttributes = attributeMaps.bindState.get(stateKey);
 	if (!affectedAttributes) return;