+++ packages/elements/src/core/constants.ts
@@ -17,6 +17,7 @@ const C = {
 			},
 			denote: {
 				action: "@",
+				state: "$",
 			},
 		},
 	},