+++ packages/elements/src/handlers.ts
@@ -1,2 +1,3 @@
 export { default as events } from "./handlers/events/index.js";
 export { default as dom } from "./handlers/dom/index.js";
+export { default as trap } from "./handlers/trap/index.js";