+++ packages/elements/src/helpers/parse-bind-value.ts
@@ -1,8 +1,12 @@
+import C from "../core/constants.js";
+import { extractBaseStateKey } from "../helpers.js";
+
 /**
  * Determines if the value is an action or state and returns the value
  * - If the attribute value is a dot notation, return the first part
  * - If the attribute value is a bracket notation, return the first part
  * - If the value contains @, its action
+ * - If the value contains $, its state
  *
  * These are returned with the scope still attached
  */
@@ -11,31 +15,19 @@ const parseBindValue = (
 ): {
 	type: "state" | "action";
 	value: string;
-} => {
-	if (value.includes("@")) {
+} | null => {
+	if (value.includes(C.defaults.attributes.denote.action)) {
 		return { type: "action", value: value };
 	}
 
-	const dotIndex = value.indexOf(".");
-	const bracketIndex = value.indexOf("[");
-
-	if (dotIndex === -1 && bracketIndex === -1) {
-		return { type: "state", value: value };
-	}
-
-	// dot notation first - or only present. key.test[0], key.test
-	if (bracketIndex === -1 || (dotIndex !== -1 && dotIndex < bracketIndex)) {
+	if (value.includes(C.defaults.attributes.denote.state)) {
 		return {
 			type: "state",
-			value: value.split(".")[0] ?? value,
+			value: extractBaseStateKey(value),
 		};
 	}
 
-	// bracket notation first / only present. key[0].test, key[0]
-	return {
-		type: "state",
-		value: value.split("[")[0] ?? value,
-	};
+	return null;
 };
 
 export default parseBindValue;